add alert before deleting an organization

+++ src/components/organization/index.tsx
@@ -1,12 +1,11 @@
-import DeleteIcon from "@mui/icons-material/Delete";
 import FolderIcon from "@mui/icons-material/Folder";
 import Avatar from "@mui/material/Avatar";
-import IconButton from "@mui/material/IconButton";
 import List from "@mui/material/List";
 import ListItem from "@mui/material/ListItem";
 import ListItemAvatar from "@mui/material/ListItemAvatar";
 import ListItemText from "@mui/material/ListItemText";
 import * as queries from "../../generated";
+import AlertDialog from "./alert";
 
 const Organization = () => {
   const { data, error, loading } = queries.useOrganizationsQuery();
@@ -28,15 +27,13 @@ const Organization = () => {
         <ListItem
           key={organization?.id}
           secondaryAction={
-            <IconButton edge="end" aria-label="delete">
-              <DeleteIcon />
-            </IconButton>
-          }
-          onClick={() =>
-            organization &&
-            deleteOrganizationMutation({
-              variables: { id: organization.id },
-            })
+            <AlertDialog
+              action={() =>
+                deleteOrganizationMutation({
+                  variables: { id: organization?.id },
+                })
+              }
+            />
           }
         >
           <ListItemAvatar>
